u8 -> u16

+++ src/lib.rs
@@ -7,20 +7,20 @@ use size::Size;
 pub struct M {
     rows: usize,
     columns: usize,
-    values: Vec<u8>
+    values: Vec<u16>
 }
 
 pub trait Convert {
     fn into_matrix(self, s: &Size) -> M;
 }
-impl Convert for Vec<u8>{
+impl Convert for Vec<u16>{
     fn into_matrix(self, s: &Size) -> M {
-        let mut v: Vec<u8> = Vec::new();
+        let mut v: Vec<u16> = Vec::new();
         let count = s.rows() * s.columns();
         let mut it = self.iter();
         let mut idx = 0;
         while idx < count {
-            let value: u8 = match it.next() {
+            let value: u16 = match it.next() {
                 Some(v) => v.clone(),
                 None => 0
             };
@@ -32,29 +32,29 @@ impl Convert for Vec<u8>{
 }
 
 pub trait Matrix {
-    fn set(&mut self, p: &Position, v: u8);
-    fn get(&self, p: &Position) -> u8;
+    fn set(&mut self, p: &Position, v: u16);
+    fn get(&self, p: &Position) -> u16;
     fn sum(&self, m: &M) -> M;
     fn transpose(&self) -> M;
 }
 
 impl Matrix for M {
-    fn set(&mut self, p: &Position, v: u8) {
+    fn set(&mut self, p: &Position, v: u16) {
         self.values[p.column() + (p.row() * self.columns)] = v;
     }
-    fn get(&self, p: &Position) -> u8 {
+    fn get(&self, p: &Position) -> u16 {
         self.values[p.column() + (p.row() * self.columns)]
     }
     fn sum(&self, m: &M) -> M {
         assert!((self.rows == m.rows) & (self.columns == m.columns));
         let zv = self.values.iter().zip(&m.values);
-        let v: Vec<u8> = zv.map(|(a, b)| {
+        let v: Vec<u16> = zv.map(|(a, b)| {
             a + b
         }).collect();
         M { rows: self.rows, columns: self.columns, values: v}
     }
     fn transpose(&self) -> M {
-        let mut v: Vec<u8> = Vec::new();
+        let mut v: Vec<u16> = Vec::new();
         for column in 0..self.columns {
             for row in 0..self.rows {
                 let value = self.get(&(row, column));
